Added gold counter as well as swapped int to BigInteger to prevent overflow

+++ Assets/Scripts/Backend/Player Saves/PlayerData.cs
@@ -1,10 +1,11 @@
 using System.Collections.Generic;
+using System.Numerics;
 
 [System.Serializable]
 public class PlayerData
 {
     //Game Stats
-    public int PlayerGold;
+    public BigInteger PlayerGold;
     public Dictionary<int, int> PlayerStatUpgradeDictionary;
 
     //Settings
